fix: Converted the getFlagEmoji util function into a reusable component and consolidated all related logic into it.
